Add test case for a ledger account object (RED)
This test establishes a ledger account with some sample income and
expense posts, then checks that the subtotals and totals come out
properly.

Due to a limitation in Ledger, we can't initiate separate ledger objects
for each test case, so we create a single (read-only) ledger object to
share across all of the tests.